+++ src/main/java/doore/study/application/CurriculumItemCommandService.java
@@ -7,11 +7,11 @@ import doore.member.domain.Participant;
 import doore.member.domain.repository.ParticipantRepository;
 import doore.study.application.dto.request.CurriculumItemRequest;
 import doore.study.domain.CurriculumItem;
-import doore.study.domain.CurriculumItemRepository;
+import doore.study.domain.repository.CurriculumItemRepository;
 import doore.study.domain.ParticipantCurriculumItem;
-import doore.study.domain.ParticipantCurriculumItemRepository;
+import doore.study.domain.repository.ParticipantCurriculumItemRepository;
 import doore.study.domain.Study;
-import doore.study.domain.StudyRepository;
+import doore.study.domain.repository.StudyRepository;
 import doore.study.exception.CurriculumItemException;
 import doore.study.exception.StudyException;
 import java.util.List;
