+++ src/main/java/doore/study/application/StudyCommandService.java
@@ -7,9 +7,9 @@ import doore.study.application.dto.request.CurriculumItemRequest;
 import doore.study.application.dto.request.StudyCreateRequest;
 import doore.study.application.dto.request.StudyUpdateRequest;
 import doore.study.domain.CurriculumItem;
-import doore.study.domain.CurriculumItemRepository;
+import doore.study.domain.repository.CurriculumItemRepository;
 import doore.study.domain.Study;
-import doore.study.domain.StudyRepository;
+import doore.study.domain.repository.StudyRepository;
 import doore.study.domain.StudyStatus;
 import doore.study.exception.StudyException;
 import doore.team.domain.TeamRepository;
