+++ src/main/java/doore/study/application/StudyQueryService.java
@@ -4,7 +4,7 @@ import static doore.study.exception.StudyExceptionType.NOT_FOUND_STUDY;
 
 import doore.study.application.dto.response.StudyDetailResponse;
 import doore.study.domain.Study;
-import doore.study.domain.StudyRepository;
+import doore.study.domain.repository.StudyRepository;
 import doore.study.exception.StudyException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
