+++ src/main/java/doore/study/domain/repository/ParticipantCurriculumItemRepository.java
@@ -1,5 +1,6 @@
-package doore.study.domain;
+package doore.study.domain.repository;
 
+import doore.study.domain.ParticipantCurriculumItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ParticipantCurriculumItemRepository extends JpaRepository<ParticipantCurriculumItem, Long> {
