+++ src/main/java/doore/study/domain/repository/StudyRepository.java
@@ -1,5 +1,6 @@
-package doore.study.domain;
+package doore.study.domain.repository;
 
+import doore.study.domain.Study;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StudyRepository extends JpaRepository<Study, Long> {
