+++ src/test/java/doore/restdocs/docs/CurriculumItemApiDocsTest.java
@@ -13,8 +13,8 @@ import doore.restdocs.RestDocsTest;
 import doore.study.api.CurriculumItemController;
 import doore.study.application.CurriculumItemCommandService;
 import doore.study.application.dto.request.CurriculumItemRequest;
-import doore.study.domain.CurriculumItemRepository;
-import doore.study.domain.StudyRepository;
+import doore.study.domain.repository.CurriculumItemRepository;
+import doore.study.domain.repository.StudyRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
