+++ src/test/java/doore/study/CurriculumItemFixture.java
@@ -2,7 +2,7 @@ package doore.study;
 
 import doore.study.domain.CurriculumItem;
 import doore.study.domain.Study;
-import doore.study.domain.StudyRepository;
+import doore.study.domain.repository.StudyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
