+++ src/test/java/doore/study/api/CurriculumItemControllerTest.java
@@ -12,9 +12,9 @@ import doore.study.CurriculumItemFixture;
 import doore.study.StudyFixture;
 import doore.study.application.dto.request.CurriculumItemRequest;
 import doore.study.domain.CurriculumItem;
-import doore.study.domain.CurriculumItemRepository;
+import doore.study.domain.repository.CurriculumItemRepository;
 import doore.study.domain.Study;
-import doore.study.domain.StudyRepository;
+import doore.study.domain.repository.StudyRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
