+++ src/test/java/doore/study/api/StudyControllerTest.java
@@ -8,7 +8,7 @@ import static org.mockito.Mockito.mock;
 import doore.helper.IntegrationTest;
 import doore.study.application.dto.request.StudyCreateRequest;
 import doore.study.domain.Study;
-import doore.study.domain.StudyRepository;
+import doore.study.domain.repository.StudyRepository;
 import doore.team.domain.Team;
 import doore.team.domain.TeamRepository;
 import java.time.LocalDate;
