+++ src/test/java/doore/study/application/CurriculumItemCommandServiceTest.java
@@ -9,9 +9,9 @@ import doore.helper.IntegrationTest;
 import doore.study.StudyFixture;
 import doore.study.application.dto.request.CurriculumItemRequest;
 import doore.study.domain.CurriculumItem;
-import doore.study.domain.CurriculumItemRepository;
+import doore.study.domain.repository.CurriculumItemRepository;
 import doore.study.domain.Study;
-import doore.study.domain.StudyRepository;
+import doore.study.domain.repository.StudyRepository;
 import doore.study.exception.CurriculumItemException;
 import java.util.Arrays;
 import java.util.List;
