+++ src/test/java/doore/study/application/StudyServiceTest.java
@@ -14,7 +14,7 @@ import doore.helper.IntegrationTest;
 import doore.study.application.dto.request.StudyCreateRequest;
 import doore.study.application.dto.request.StudyUpdateRequest;
 import doore.study.domain.Study;
-import doore.study.domain.StudyRepository;
+import doore.study.domain.repository.StudyRepository;
 import doore.study.domain.StudyStatus;
 import doore.study.exception.StudyException;
 import doore.team.domain.Team;
